backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
simplemma
rapidfuzz
python-dotenv
pymongo
certifi
